+++ libs/connectors/muapi/genblaze_muapi/__init__.py
@@ -1,0 +1,6 @@
+"""MuAPI media provider for genblaze."""
+
+from ._version import __version__
+from .provider import MuAPIProvider
+
+__all__ = ["MuAPIProvider", "__version__"]